Statement wrapper with input parameter binding.
This adds a wrapper around MYSQL_STMT (prepared statements), with support
for easy-to-use binding of input parameters (supporting int64_t and
std::string for TEXT/BLOB as well as NULL for now).

Handling of the memory (keeping copies of the data as needed) and filling
of the MYSQL_BIND structs is done behind the scenes.