Support options in URL parsing.
This extends MySQL URL parsing to support passing options (key/value
pairs) in the URL, of the form:

  mysql://user:password@host:port/db?foo=bar&baz=42

We will use that to pass data about client certificates (but the URL parsing
is universal).

As part of this change, the ParseUrl function is replaced by a UrlParser
class, which holds the parts parsed by itself, so callers don't have to
declare more and more values explicitly (and can then just use getters
to extract the bits they need).